Fix flow and reorganise typings
